Get rid of old Install/Upgrade scripts, migrate to declarative schema completely. (#159)
(cherry picked from commit 817c32aee39e6700483b3a3d3bf409a033361674)
