--- Setup/InstallSchema.php
@@ -1,51 +1,0 @@
-<?php
-/**
- * Copyright © 2016 Magento. All rights reserved.
- * See COPYING.txt for license details.
- */
-
-namespace Klaviyo\Reclaim\Setup;
-use Magento\Framework\Setup\InstallSchemaInterface;
-use Magento\Framework\Setup\ModuleContextInterface;
-use Magento\Framework\Setup\SchemaSetupInterface;
-
-/**
- * @codeCoverageIgnore
- */
-class InstallSchema implements InstallSchemaInterface
-{
-    /**
-     * {@inheritdoc}
-     * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
-     */
-    public function install(SchemaSetupInterface $setup, ModuleContextInterface $context)
-    {
-        $installer = $setup;
-        $installer->startSetup();
-
-        $eavTable1 = $installer->getTable('quote');
-        $eavTable2 = $installer->getTable('sales_order');
-
-        $columns = [
-            'kl_sms_consent' => [
-                'type' => \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                'nullable' => true,
-                'comment' => 'SMS Consent'
-            ],
-            'kl_email_consent' => [
-                'type' => \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                'nullable' => true,
-                'comment' => 'Email Consent'
-            ],
-        ];
-
-        $connection = $installer->getConnection();
-
-        foreach($columns as $name => $definition) {
-            $connection->addColumn($eavTable1, $name, $definition);
-            $connection->addColumn($eavTable2, $name, $definition);
-        }
-
-        $installer->endSetup();
-    }
-}
